# refactor(response_cache): do not serialize in json before hashing, and improve hashing