--- a/apollo-router/src/plugins/response_cache/mod.rs
+++ b/apollo-router/src/plugins/response_cache/mod.rs
@@ -4,6 +4,7 @@ pub(crate) mod invalidation_endpoint;
 pub(crate) mod metrics;
 pub(crate) mod plugin;
 pub(crate) mod postgres;
+pub(crate) mod serde_blake3;
 #[cfg(all(
     test,
     any(not(feature = "ci"), all(target_arch = "x86_64", target_os = "linux"))