Refactor response caching hashes for primary cache keys (both for entities and root fields). It also switches the hash from sha256 to blake3 because it's faster and safer. 

Benchmarks:

+ Before:
```
Requests      [total, rate, throughput]         75000, 250.00, 249.98
Duration      [total, attack, wait]             5m0s, 5m0s, 24.589ms
Latencies     [min, mean, 50, 90, 95, 99, max]  12.633ms, 20.941ms, 20.655ms, 24.573ms, 25.539ms, 30.076ms, 211.837ms
Bytes In      [total, mean]                     120450000, 1606.00
Bytes Out     [total, mean]                     5475000, 73.00
Success       [ratio]                           100.00%
Status Codes  [code:count]                      200:75000
```

+ After:
```
Requests      [total, rate, throughput]         75000, 250.00, 249.99
Duration      [total, attack, wait]             5m0s, 5m0s, 20.946ms
Latencies     [min, mean, 50, 90, 95, 99, max]  10.175ms, 16.761ms, 16.428ms, 20.03ms, 20.436ms, 28.938ms, 142.605ms
Bytes In      [total, mean]                     120450000, 1606.00
Bytes Out     [total, mean]                     5475000, 73.00
Success       [ratio]                           100.00%
Status Codes  [code:count]                      200:75000 
```

<!-- [ROUTER-1161] -->
<!-- [ROUTER-1159] -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR is marked ready-for-review.

- [x] PR description explains the motivation for the change and relevant context for reviewing
- [x] PR description links appropriate GitHub/Jira tickets (creating when necessary)
- [ ] Changeset is included for user-facing changes
- [ ] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [x] Performance impact assessed and acceptable
- [ ] Metrics and logs are added[^3] and documented
- Tests added and passing[^4]
    - [x] Unit tests
    - [x] Integration tests
    - [x] Manual tests, as necessary

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention of other (impacted) groups. Please endeavour to do this before seeking PR approval. The mechanism for doing this will vary considerably, so use your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where applicable please try to document configuration examples.
[^3]: A lot of (if not most) features benefit from built-in observability and `debug`-level logs. Please read [this guidance](https://github.com/apollographql/router/blob/dev/dev-docs/metrics.md#adding-new-metrics) on metrics best-practices.
[^4]: Tick whichever testing boxes are applicable. If you are adding Manual Tests, please document the manual testing (extensively) in the Exceptions.


[ROUTER-1161]: https://apollographql.atlassian.net/browse/ROUTER-1161?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[ROUTER-1159]: https://apollographql.atlassian.net/browse/ROUTER-1159?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ